+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Added ConfigUtils for accessing orphaned entries  
 - Added PersistentItemReference and ItemUtils for referencing items in configs and in gameplay systems   
 - Added SpawnableScrapUtils for tracking currently spawnable scrap  
-- Added LCUtils_Preloader for patching in needed Unity MonoBehaviour methods (e.g. Awake(), Start(), etc.)  
+- Added NumberExtensions for non-finite number handling  
   
 ## v1.0.5